+++ examples/resources/meshstack_tag_definition/import.sh
@@ -1,0 +1,5 @@
+# import via tag definition name
+terraform import 'meshstack_tag_definition.example' 'target_kind.key'
+
+# import the example tag definition
+terraform import 'meshstack_tag_definition.example' 'meshProject.example-key'