+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 FEATURES:
 - Added `meshstack_tag_definitions` data source.
+- Added `meshstack_tag_definition` data source.
+- Added `meshstack_tag_definition` resource.
 
 ## v0.5.5
 
 FIXES:
 - HTTP response code for tenant creation is now 201.
 - HTTP response code for project creation is now 201.
-
 
 ## v0.5.4